+++ src/config_pressuremon.cpp
@@ -47,6 +47,24 @@ void PressuremonConfig::createJson(JsonObject& doc) const {
       serialized(String(getPressureSensorCorrection(0), DECIMALS_PRESSURE));
   doc[CONFIG_PRESSURE1_ADJUSTMENT] =
       serialized(String(getPressureSensorCorrection(1), DECIMALS_PRESSURE));
+
+  doc[CONFIG_CUSTOM_MIN_VOLTAGE] =
+      serialized(String(getCustomAnalogMinV(0), 3));
+  doc[CONFIG_CUSTOM_MAX_VOLTAGE] =
+      serialized(String(getCustomAnalogMaxV(0), 3));
+  doc[CONFIG_CUSTOM_MIN_PRESSURE] =
+      serialized(String(getCustomAnalogMinKpa(0), DECIMALS_PRESSURE));
+  doc[CONFIG_CUSTOM_MAX_PRESSURE] =
+      serialized(String(getCustomAnalogMaxKpa(0), DECIMALS_PRESSURE));
+
+  doc[CONFIG_CUSTOM1_MIN_VOLTAGE] =
+      serialized(String(getCustomAnalogMinV(1), 3));
+  doc[CONFIG_CUSTOM1_MAX_VOLTAGE] =
+      serialized(String(getCustomAnalogMaxV(1), 3));
+  doc[CONFIG_CUSTOM1_MIN_PRESSURE] =
+      serialized(String(getCustomAnalogMinKpa(1), DECIMALS_PRESSURE));
+  doc[CONFIG_CUSTOM1_MAX_PRESSURE] =
+      serialized(String(getCustomAnalogMaxKpa(1), DECIMALS_PRESSURE));
 }
 
 void PressuremonConfig::parseJson(JsonObject& doc) {
@@ -71,6 +89,24 @@ void PressuremonConfig::parseJson(JsonObject& doc) {
   if (!doc[CONFIG_PRESSURE1_ADJUSTMENT].isNull())
     setPressureSensorCorrection(doc[CONFIG_PRESSURE1_ADJUSTMENT].as<float>(),
                                 1);
+
+  if (!doc[CONFIG_CUSTOM_MIN_VOLTAGE].isNull())
+    setCustomAnalogMinV(doc[CONFIG_CUSTOM_MIN_VOLTAGE].as<float>(), 0);
+  if (!doc[CONFIG_CUSTOM_MAX_VOLTAGE].isNull())
+    setCustomAnalogMaxV(doc[CONFIG_CUSTOM_MAX_VOLTAGE].as<float>(), 0);
+  if (!doc[CONFIG_CUSTOM_MIN_PRESSURE].isNull())
+    setCustomAnalogMinKpa(doc[CONFIG_CUSTOM_MIN_PRESSURE].as<float>(), 0);
+  if (!doc[CONFIG_CUSTOM_MAX_PRESSURE].isNull())
+    setCustomAnalogMaxKpa(doc[CONFIG_CUSTOM_MAX_PRESSURE].as<float>(), 0);
+
+  if (!doc[CONFIG_CUSTOM1_MIN_VOLTAGE].isNull())
+    setCustomAnalogMinV(doc[CONFIG_CUSTOM1_MIN_VOLTAGE].as<float>(), 1);
+  if (!doc[CONFIG_CUSTOM1_MAX_VOLTAGE].isNull())
+    setCustomAnalogMaxV(doc[CONFIG_CUSTOM1_MAX_VOLTAGE].as<float>(), 1);
+  if (!doc[CONFIG_CUSTOM1_MIN_PRESSURE].isNull())
+    setCustomAnalogMinKpa(doc[CONFIG_CUSTOM1_MIN_PRESSURE].as<float>(), 1);
+  if (!doc[CONFIG_CUSTOM1_MAX_PRESSURE].isNull())
+    setCustomAnalogMaxKpa(doc[CONFIG_CUSTOM1_MAX_PRESSURE].as<float>(), 1);
 }
 
 #endif  // PRESSUREMON