+++ src/pressure.hpp
@@ -27,6 +27,7 @@ SOFTWARE.
 #if defined(PRESSUREMON)
 
 #include <Arduino.h>
+#include <SoftWire.h>
 #include <Wire.h>
 
 #include <log.hpp>
@@ -50,6 +51,7 @@ enum PressureSensorType {
   SensorXidibeiXDB401_I2C_KPa_3500 = 13,  // 0-3.5 MPa
   SensorXidibeiXDB401_I2C_KPa_4000 = 14,  // 0-4 MPa
 
+  /*
   SensorXidibeiXDB401_Analog_KPa_200 = 101,   // 0-0.2 MPa
   SensorXidibeiXDB401_Analog_KPa_400 = 102,   // 0-0.4 MPa
   SensorXidibeiXDB401_Analog_KPa_500 = 103,   // 0-0.5 MPa
@@ -64,6 +66,9 @@ enum PressureSensorType {
   SensorXidibeiXDB401_Analog_KPa_3000 = 112,  // 0-3 MPa
   SensorXidibeiXDB401_Analog_KPa_3500 = 113,  // 0-3.5 MPa
   SensorXidibeiXDB401_Analog_KPa_4000 = 114,  // 0-4 MPa
+  */
+
+  SensorCustom_Analog = 1000,  // Custom values
 };
 
 // Inteface towards config class for pressure sensor related settings
@@ -84,6 +89,11 @@ class PressureConfigInterface {
   virtual bool saveFile() = 0;
 
   virtual PressureSensorType getPressureSensorType(int idx) const = 0;
+
+  virtual float getCustomAnalogMinV(int idx) const = 0;
+  virtual float getCustomAnalogMaxV(int idx) const = 0;
+  virtual float getCustomAnalogMinKpa(int idx) const = 0;
+  virtual float getCustomAnalogMaxKpa(int idx) const = 0;
 };
 
 class PressureSensorInterface {
@@ -115,7 +125,7 @@ class PressureSensor {
     _pressureConfig = pressureConfig;
   }
 
-  void setup(uint8_t idx, TwoWire* wire);
+  void setup(uint8_t idx, TwoWire* wire, SoftWire* softWire = nullptr);
 
   bool read();
   bool isActive();
@@ -134,7 +144,9 @@ class PressureSensor {
 };
 
 extern PressureSensor myPressureSensor;
-// extern PressureSensor myPressureSensor1;
+#if defined(ENABLE_SECOND_SENSOR)
+extern PressureSensor myPressureSensor1;
+#endif
 
 #endif  // PRESSUREMON